Move models to a separate file
